Fix to stop toot status line overflowing
Added a customizable option `mastodon-alt-tl-toot-status-align-space`, default value 2, to avoid having to hard-code value in mastodon-alt.el. Setting this to a larger integer (5 or larger, 8 works for me) prevents the status line (i.e. the bottom-most part of the toot containing the number of Favorites, Boosts, Replies, and Bookmarks) from overflowing.